schitt-creek-apoth/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.8.0
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
crispy-bootstrap5==2024.2
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.1
django-ajax-datatable==4.5.0
django-appconf==1.0.6
django-compressor==4.5.1
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-datatable==0.3.1
django-datatable-view==2.1.6
django-dynamic-datatb==1.0.25
django-filter==24.3
django-htmx==1.19.0
django-js-lib-ag-grid-community==25.3.0.0
django-seed==0.3.1
django-tables2==2.7.0
djangorestframework==3.15.2
djangorestframework-datatables==0.7.2
docopt==0.6.2
environs==11.0.0
executing==2.0.1
Faker==27.0.0
fastjsonschema==2.20.0
filelock==3.15.4
fonttools==4.53.1
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nodeenv==1.9.1
nose==1.3.7
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
parts==1.7.0
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pip-system-certs==4.0
pipenv==2024.0.1
pipreqs==0.5.0
platformdirs==4.2.2
prettier-printer==0.1.4
products==1.2.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pyparsing==3.1.2
pytailwindcss==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.1.1
rcssmin==1.1.2
referencing==0.35.1
requests==2.32.3
rjsmin==1.2.2
rpds-py==0.20.0
setuptools==72.2.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.3.0
toposort==1.10
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
yarg==0.1.9
